Dequeuing video buffer stalls.

+++ include/zephyr/drivers/video.h
@@ -150,7 +150,7 @@ struct video_buffer {
 	/** type of the buffer */
 	enum video_buf_type type;
 	/** pointer to the start of the buffer. */
-	uint8_t *buffer;
+	uint32_t *buffer;
 	/** index of the buffer, optionally set by the application */
 	uint8_t index;
 	/** size of the buffer in bytes. */
